change default pub page
